+++ src/LlmTornado.Agents.Samples/CodingAgent/CSCodingAgent.cs
@@ -3,6 +3,8 @@ using LlmTornado.Agents.ChatRuntime;
 using LlmTornado.Agents.ChatRuntime.Orchestration;
 using LlmTornado.Agents.ChatRuntime.RuntimeConfigurations;
 using LlmTornado.Agents.DataModels;
+using LlmTornado.Agents.Samples.Common;
+using LlmTornado.Agents.Samples.DataModels;
 using LlmTornado.Chat;
 using LlmTornado.Chat.Models;
 using System;
@@ -14,298 +16,9 @@ using System.Linq;
 using System.Net.Sockets;
 using System.Text;
 using System.Threading.Tasks;
-using static LlmTornado.Demo.ExampleAgents.CSCodingAgent.CodeUtility;
 
-namespace LlmTornado.Demo.ExampleAgents.CSCodingAgent;
+namespace LlmTornado.Agents.Samples.CodingAgent;
 
-#region Data Models
-public struct TaskRequest
-{
-    public string Task { get; set; }
-    public string Context { get; set; }
-    public TaskRequest(string task, string context = "")
-    {
-        Task = task;
-        Context = context;
-    }
-}
-
-public struct TaskPlan
-{
-    public string OriginalTask { get; set; }
-    public string[] RequiredAgents { get; set; }
-    public string ExecutionPlan { get; set; }
-    public string[] StepDescriptions { get; set; }
-    public TaskPlan(string originalTask, string[] requiredAgents, string executionPlan, string[] stepDescriptions)
-    {
-        OriginalTask = originalTask;
-        RequiredAgents = requiredAgents;
-        ExecutionPlan = executionPlan;
-        StepDescriptions = stepDescriptions;
-    }
-}
-
-public struct AgentAction
-{
-    public string AgentType { get; set; }
-    public string Action { get; set; }
-    public string Reasoning { get; set; }
-    public string Parameters { get; set; }
-    public AgentAction(string agentType, string action, string reasoning, string parameters = "")
-    {
-        AgentType = agentType;
-        Action = action;
-        Reasoning = reasoning;
-        Parameters = parameters;
-    }
-}
-
-public struct WebSearchResult
-{
-    public string Query { get; set; }
-    public string Results { get; set; }
-    public string Summary { get; set; }
-    public WebSearchResult(string query, string results, string summary)
-    {
-        Query = query;
-        Results = results;
-        Summary = summary;
-    }
-}
-
-public struct FileOperationResult
-{
-    public string Operation { get; set; }
-    public string FilePath { get; set; }
-    public string Content { get; set; }
-    public bool Success { get; set; }
-    public string ErrorMessage { get; set; }
-    public FileOperationResult(string operation, string filePath, string content, bool success, string errorMessage = "")
-    {
-        Operation = operation;
-        FilePath = filePath;
-        Content = content;
-        Success = success;
-        ErrorMessage = errorMessage;
-    }
-}
-
-public struct CodeExecutionResult
-{
-    public string Code { get; set; }
-    public string Language { get; set; }
-    public string Output { get; set; }
-    public bool Success { get; set; }
-    public string ErrorMessage { get; set; }
-    public CodeExecutionResult(string code, string language, string output, bool success, string errorMessage = "")
-    {
-        Code = code;
-        Language = language;
-        Output = output;
-        Success = success;
-        ErrorMessage = errorMessage;
-    }
-}
-
-public struct TerminalCommandResult
-{
-    public string Command { get; set; }
-    public string Output { get; set; }
-    public int ExitCode { get; set; }
-    public bool Success { get; set; }
-    public TerminalCommandResult(string command, string output, int exitCode, bool success)
-    {
-        Command = command;
-        Output = output;
-        ExitCode = exitCode;
-        Success = success;
-    }
-}
-
-public struct AgentExecutionResults
-{
-    public string OriginalTask { get; set; }
-    public string WebSearchResults { get; set; }
-    public string FileOperationResults { get; set; }
-    public string CodeExecutionResults { get; set; }
-    public string TerminalResults { get; set; }
-    public string[] ActionsPerformed { get; set; }
-    public AgentExecutionResults(string originalTask, string webResults, string fileResults, string codeResults, string terminalResults, string[] actionsPerformed)
-    {
-        OriginalTask = originalTask;
-        WebSearchResults = webResults;
-        FileOperationResults = fileResults;
-        CodeExecutionResults = codeResults;
-        TerminalResults = terminalResults;
-        ActionsPerformed = actionsPerformed;
-    }
-}
-public struct ProgramResult
-{
-    public CodeItem[] items { get; set; }
-    public ProgramResult(CodeItem[] items)
-    {
-        this.items = items;
-    }
-}
-
-public struct CodeItem
-{
-    public string filePath { get; set; }
-    public string code { get; set; }
-
-    public CodeItem(string path, string code)
-    {
-        filePath = path;
-        this.code = code;
-    }
-}
-
-public struct ProgramResultOutput
-{
-    public ProgramResult Result { get; set; }
-    public string ProgramRequest { get; set; }
-    public ProgramResultOutput(ProgramResult result, string request)
-    {
-        Result = result;
-        ProgramRequest = request;
-    }
-}
-
-public struct CodeBuildInfoOutput
-{
-    public CodeBuildInfo BuildInfo { get; set; }
-    public ProgramResultOutput ProgramResult { get; set; }
-
-    public CodeBuildInfoOutput() { }
-    public CodeBuildInfoOutput(CodeBuildInfo info, ProgramResultOutput codeResult)
-    {
-        BuildInfo = info;
-        ProgramResult = codeResult;
-    }
-}
-
-public struct CodeReview
-{
-    public string ReviewSummary { get; set; }
-    public CodeReviewItem[] Items { get; set; }
-
-    public CodeReview(CodeReviewItem[] item)
-    {
-        Items = item;
-    }
-    public override string ToString()
-    {
-        return $""""
-            From Code Review Summary:
-            {ReviewSummary}
-
-            Items to fix:
-
-            {string.Join("\n\n", Items)}
-
-            """";
-    }
-}
-
-public struct CodeReviewItem
-{
-    public string CodePath { get; set; }
-    public string CodeError { get; set; }
-    public string SuggestedFix { get; set; }
-    public CodeReviewItem(string codePath, string codeError, string suggestedFix)
-    {
-        CodePath = codePath;
-        CodeError = codeError;
-        SuggestedFix = suggestedFix;
-    }
-
-    public override string ToString()
-    {
-        return $"""
-
-             File: {CodePath}
-             
-             Had Error: 
-             {CodeError}
-
-             Suggested Fix:
-             {SuggestedFix}
-
-             """;
-    }
-}
-
-public class CodeBuildInfo
-{
-
-    public string BuildPath { get; set; }
-
-    public string ProjectName { get; set; }
-    public ExecutableOutputResult? ExecutableResult { get; set; }
-
-    public BuildOutputResult? BuildResult { get; set; }
-
-    public CodeBuildInfo() { }
-
-    public CodeBuildInfo(string buildPath, string projectName)
-    {
-        BuildPath = buildPath;
-        ProjectName = projectName;
-    }
-}
-public class ExecutableOutputResult
-{
-    private bool executionCompleted = false;
-
-    public string Output { get; set; }
-    public string Error { get; set; }
-    public bool ExecutionCompleted { get => executionCompleted; set => executionCompleted = value; }
-
-    public ExecutableOutputResult() { }
-
-    public ExecutableOutputResult(string output, string error, bool completed)
-    {
-        Output = output;
-        Error = error;
-        ExecutionCompleted = completed;
-    }
-}
-#endregion
-
-public class CoderRunnable : OrchestrationRunnable<TaskPlan, AgentExecutionResults>
-{
-    TornadoApi Client { get; set; }
-    public CoderRunnable(TornadoApi client, Orchestration orchestrator) : base(orchestrator)
-    {
-        Client = client;
-    }
-
-    public override async ValueTask<AgentExecutionResults> Invoke(RunnableProcess<TaskPlan, AgentExecutionResults> process)
-    {
-        string codePrompt = $"""
-            Original Task: {process.Input.OriginalTask}
-            
-            Execution Plan: {process.Input.ExecutionPlan}
-            
-            As the Coder agent, write, review, or execute code as needed for this task. Provide comprehensive programming solutions that will be used by the Orchestrator to create the final deliverable.
-            """;
-
-        CodingAgentConfiguration codingAgentConfiguration = new CodingAgentConfiguration(Client);
-        ChatMessage msg = await codingAgentConfiguration.AddToChatAsync(new ChatMessage(Code.ChatMessageRoles.User, codePrompt));
-
-        string codeResults = msg.Content ?? "";
-
-        return new AgentExecutionResults(
-            process.Input.OriginalTask,
-            "",
-            "",
-            codeResults,
-            "",
-            new[] { "Coder: Provided programming and coding solutions" }
-        );
-    }
-}
 
 public class CodingAgentConfiguration : OrchestrationRuntimeConfiguration
 {
@@ -344,7 +57,7 @@ public class CodingAgentConfiguration : OrchestrationRuntimeConfiguration
 
     public override void OnRuntimeInitialized()
     {
-        CreateNewProject(ProjectName);
+        CodeUtility.CreateNewProject(ProjectName);
 
         Console.WriteLine($"Created new project at {FileIOUtility.SafeWorkingDirectory}");
     }
@@ -460,7 +173,7 @@ public class CodingAgentConfiguration : OrchestrationRuntimeConfiguration
 
             //build the project code
             //In theory here i could setup a lot of different code to build
-            CodeBuildInfo codeInfo = BuildAndRunProject(FileIOUtility.SafeWorkingDirectory, "netcoreapp3.1");
+            CodeBuildInfo codeInfo = CodeUtility.BuildAndRunProject(FileIOUtility.SafeWorkingDirectory, "netcoreapp3.1");
 
             //Report the results of the build
             return ValueTask.FromResult(new CodeBuildInfoOutput(codeInfo, programResult.Input));
@@ -549,328 +262,4 @@ public class CodingAgentConfiguration : OrchestrationRuntimeConfiguration
 
     }
 }
-
-public class FileIOUtility
-{
-    public static string? SafeWorkingDirectory { get; set; }
-
-    public static string ReadFile(string filePath)
-    {
-        try
-        {
-            if (string.IsNullOrEmpty(FileIOUtility.SafeWorkingDirectory))
-            {
-                throw new InvalidOperationException("SafeWorkingDirectory is not set. Please set SafeWorkingDirectory before reading directories.");
-            }
-
-            filePath = filePath.Trim();
-
-            Path.GetInvalidFileNameChars().ToList().ForEach(c =>
-            {
-                if (Path.GetFileName(filePath).Contains(c))
-                {
-                    throw new ArgumentException($"File name cannot contain {c}");
-                }
-            }
-            );
-            // Validate the file path
-            if (string.IsNullOrEmpty(filePath))
-            {
-                throw new ArgumentException("File path cannot be null or empty.");
-            }
-            if (filePath.StartsWith("..") || filePath.StartsWith("/"))
-            {
-                throw new ArgumentException("File path cannot contain relative paths like '..', or '/'.");
-            }
-            if (filePath.StartsWith("\\"))
-            {
-                filePath = filePath.TrimStart('\\');
-            }
-            string fixedPath = Path.Combine(FileIOUtility.SafeWorkingDirectory, filePath.Trim());
-
-            return File.ReadAllText(fixedPath);
-        }
-        catch (Exception ex)
-        {
-            return $"Error reading file -> {ex.Message}"; // Return empty string or handle as needed
-        }
-    }
-
-    public static void WriteFile(string filePath, string content)
-    {
-
-        if (string.IsNullOrEmpty(SafeWorkingDirectory))
-        {
-            throw new InvalidOperationException("SafeWorkingDirectory is not set. Please set SafeWorkingDirectory before reading directories.");
-        }
-
-        string FixedPath = Path.Combine(SafeWorkingDirectory, filePath);
-
-        string? directoryPath = Path.GetDirectoryName(FixedPath);
-
-        if (!string.IsNullOrEmpty(directoryPath) && !Directory.Exists(directoryPath))
-        {
-            Directory.CreateDirectory(directoryPath);
-        }
-
-        File.WriteAllText(FixedPath, content);
-    }
-
-    public static string GetAllPaths(string directory)
-    {
-        if (string.IsNullOrEmpty(SafeWorkingDirectory))
-        {
-            throw new InvalidOperationException("SafeWorkingDirectory is not set. Please set SafeWorkingDirectory before reading directories.");
-        }
-
-        List<string> allPaths = new List<string>();
-
-        GetPaths(directory, allPaths);
-
-        for (int i = 0; i < allPaths.Count; i++)
-        {
-            allPaths[i] = allPaths[i].Replace(SafeWorkingDirectory, "");
-        }
-
-        foreach (string path in allPaths)
-        {
-            Console.WriteLine(path);
-        }
-
-        return string.Join(Environment.NewLine, allPaths);
-    }
-
-    public static void GetPaths(string directory, List<string> paths)
-    {
-        if (string.IsNullOrEmpty(SafeWorkingDirectory))
-        {
-            throw new InvalidOperationException("SafeWorkingDirectory is not set. Please set SafeWorkingDirectory before reading directories.");
-        }
-
-        string SafePath = Path.Combine(SafeWorkingDirectory, directory);
-
-        try
-        {
-            // Add files in the current directory
-            paths.AddRange(Directory.GetFiles(SafePath));
-
-            // Add subdirectories in the current directory and recurse
-            string[] subdirectories = Directory.GetDirectories(SafePath);
-            //paths.AddRange(subdirectories);
-            foreach (string subdirectory in subdirectories)
-            {
-                if (subdirectory.Contains("bin")) continue;
-                if (subdirectory.Contains("obj")) continue;
-                GetPaths(subdirectory, paths);
-            }
-        }
-        catch (UnauthorizedAccessException ex)
-        {
-            Console.WriteLine($"Access denied to: {ex.Message}"); // Handle exceptions gracefully
-        }
-        catch (DirectoryNotFoundException ex)
-        {
-            Console.WriteLine($"Directory not found: {ex.Message}"); // Handle exceptions gracefully
-        }
-    }
-}
-
-public class CodeUtility
-{
-    public static CodeBuildInfo BuildAndRunProject(string pathToSolution, string? framework = "", bool runProject = false)
-    {
-        CodeBuildInfo codeBuildInfo = new CodeBuildInfo(pathToSolution, "FunctionApplication");
-        // Path to the target project's directory
-        string projectPath = pathToSolution;
-
-        // ... (Code for programmatically building the project as shown in the previous example) ...
-
-        // After the build is successful:
-        codeBuildInfo.BuildResult = BuildProject(FileIOUtility.SafeWorkingDirectory);
-
-        if (codeBuildInfo.BuildResult.BuildCompleted)
-        {
-            Console.WriteLine("Build successful! Now running the built project...");
-            // Find the executable file
-            string executablePath = FindExecutable(projectPath, "FunctionApplication", framework);
-
-            if (!string.IsNullOrEmpty(executablePath))
-            {
-                if (runProject)
-                {
-                    // Run the executable and capture its output
-                    codeBuildInfo.ExecutableResult = RunExecutableAndCaptureOutput(executablePath);
-                }
-            }
-            else
-            {
-                Console.WriteLine("Could not find the executable file.");
-            }
-        }
-        else
-        {
-            Console.WriteLine("Build failed. Cannot run the built project.");
-        }
-
-        return codeBuildInfo;
-    }
-
-    // Function to find the executable file after build
-    public static string FindExecutable(string projectPath, string projectName, string framework)
-    {
-        // Assuming a typical .NET Core/5+ project structure
-        // You might need to adjust this based on your project setup
-        string binPath = Path.Combine(projectPath, projectName, "bin", "Debug", framework); // Adjust framework if needed
-        string executableName = $"{projectName}.exe"; // Replace with your executable name
-        string executablePath = Path.Combine(binPath, executableName);
-
-        if (File.Exists(executablePath))
-        {
-            return executablePath;
-        }
-
-        // Check the Release folder as well
-        binPath = Path.Combine(projectPath, projectName, "bin", "Release", framework); // Adjust framework if needed
-        executablePath = Path.Combine(binPath, executableName);
-
-        if (File.Exists(executablePath))
-        {
-            return executablePath;
-        }
-
-        return null; // Executable not found
-    }
-
-
-    // Function to run the executable and capture its output
-    public static ExecutableOutputResult RunExecutableAndCaptureOutput(string executablePath, string? arguments = "")
-    {
-        Process process = new Process();
-        process.StartInfo.FileName = executablePath;
-        process.StartInfo.Arguments = arguments; // Pass any arguments to the executable here
-        process.StartInfo.UseShellExecute = false;
-        process.StartInfo.RedirectStandardOutput = true;
-        process.StartInfo.RedirectStandardError = true;
-        process.StartInfo.CreateNoWindow = true;
-
-        ExecutableOutputResult result = new ExecutableOutputResult();
-
-        try
-        {
-            process.Start();
-
-            // Read the output and error streams
-            result.Output = process.StandardOutput.ReadToEnd();
-            result.Error = process.StandardError.ReadToEnd();
-
-            process.WaitForExit();
-
-            if (process.ExitCode == 0)
-            {
-                Console.WriteLine("Executable Output:");
-                Console.WriteLine(result.Output);
-            }
-            else
-            {
-                Console.WriteLine("Executable Error:");
-                Console.WriteLine(result.Error);
-            }
-            result.ExecutionCompleted = true;
-            return result;
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"Error running executable: {ex.Message}");
-            result.ExecutionCompleted = false;
-            return result;
-        }
-    }
-
-    public class BuildOutputResult
-    {
-        public string Output { get; set; }
-        public string Error { get; set; }
-        public bool BuildCompleted { get; set; }
-
-        public BuildOutputResult() { }
-
-        public BuildOutputResult(string output, string error, bool completed)
-        {
-            Output = output;
-            Error = error;
-            BuildCompleted = completed;
-        }
-    }
-
-    public static bool CreateNewProject(string projectName)
-    {
-        string zipPath = "Static//Files//FunctionApplicationTemplate.zip";
-        string extractPath = FileIOUtility.SafeWorkingDirectory + "_TEMP";
-        string finalPath = FileIOUtility.SafeWorkingDirectory;
-
-        if (Directory.Exists(extractPath) || Directory.Exists(finalPath))
-        {
-            return false;
-        }
-
-        ZipFile.ExtractToDirectory(zipPath, extractPath); //Extract zip
-        Directory.Move(Path.Combine(extractPath, "FunctionApplication"), finalPath); // Rename folder
-        Directory.Delete(extractPath, true); //Delete temp folder
-
-        return true;
-    }
-
-    // Function to build the project using dotnet CLI
-    public static BuildOutputResult BuildProject(string path)
-    {
-        // Path to the target project's directory
-        string projectPath = path;
-
-        // Create a new process to run the dotnet build command
-        Process process = new Process();
-        process.StartInfo.FileName = "dotnet"; // Use "dotnet" command
-        process.StartInfo.Arguments = $"build \"{projectPath}\""; // Arguments to build the project
-        process.StartInfo.UseShellExecute = false; // Don't use the OS shell
-        process.StartInfo.RedirectStandardOutput = true; // Redirect standard output to capture build output
-        process.StartInfo.RedirectStandardError = true; // Redirect standard error to capture error messages
-        process.StartInfo.CreateNoWindow = true; // Don't create a new window for the process
-
-        BuildOutputResult result = new BuildOutputResult();
-
-        try
-        {
-            // Start the process
-            process.Start();
-
-            // Read the build output (optional)
-            result.Output = process.StandardOutput.ReadToEnd();
-            result.Error = process.StandardError.ReadToEnd();
-
-            // Wait for the process to exit
-            process.WaitForExit();
-
-            // Check the exit code to determine if the build was successful
-            if (process.ExitCode == 0)
-            {
-                Console.WriteLine("Build successful!");
-                Console.WriteLine(result.Output);
-                result.BuildCompleted = true;
-            }
-            else
-            {
-                Console.WriteLine("Build failed!");
-                Console.WriteLine(result.Error);
-                result.BuildCompleted = false;
-            }
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"Error occurred: {ex.Message}");
-            result.BuildCompleted = false;
-        }
-
-        return result;
-    }
-}
-
