+++ src/LlmTornado.Agents.Samples/DataModels/ProgramResult.cs
@@ -1,0 +1,10 @@
+﻿namespace LlmTornado.Agents.Samples.DataModels;
+public struct ProgramResult
+{
+    public CodeItem[] items { get; set; }
+    public ProgramResult(CodeItem[] items)
+    {
+        this.items = items;
+    }
+}
+
