+++ src/LlmTornado.Agents.Samples/DataModels/TaskRequest.cs
@@ -1,0 +1,13 @@
+﻿namespace LlmTornado.Agents.Samples.DataModels;
+
+public struct TaskRequest
+{
+    public string Task { get; set; }
+    public string Context { get; set; }
+    public TaskRequest(string task, string context = "")
+    {
+        Task = task;
+        Context = context;
+    }
+}
+
